+++ approvaltests/src/main/java/org/approvaltests/reporters/intellij/IntelliJPathResolver.java
@@ -1,0 +1,57 @@
+package org.approvaltests.reporters.intellij;
+
+import java.io.IOException;
+import java.nio.file.FileVisitOption;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.Comparator;
+import java.util.Objects;
+import java.util.Optional;
+
+public class IntelliJPathResolver {
+    private final String channelsPath;
+
+    public IntelliJPathResolver(Edition edition) {
+        String appData = System.getenv("LOCALAPPDATA");
+        String toolboxPath = appData + "/JetBrains/Toolbox";
+        this.channelsPath = toolboxPath + "/apps/" + edition.getDirectory() + "/ch-0/";
+    }
+
+    public String findIt() {
+        String notPresentPath = "C:\\Intelli-not-present.exe";
+        try {
+            return getIntelliJPath().map(Objects::toString).orElse(notPresentPath);
+        } catch (IOException e) {
+            return notPresentPath;
+        }
+    }
+
+    private Optional<Path> getIntelliJPath() throws IOException {
+        return Files.walk(Paths.get(channelsPath), 1, FileVisitOption.FOLLOW_LINKS)
+                .map(Path::getFileName)
+                .map(Objects::toString)
+                .filter(IntelliJCommunityReporter.Version::isVersionFile)
+                .map(IntelliJCommunityReporter.Version::new)
+                .max(Comparator.naturalOrder())
+                .map(this::getPath);
+    }
+
+    private Path getPath(IntelliJCommunityReporter.Version version) {
+        String runtimeSuffix = "/bin/idea64.exe";
+        return Paths.get(channelsPath + version.version + runtimeSuffix).toAbsolutePath();
+    }
+
+    public enum Edition {
+        Community("IDEA-C"), Ultimate("IDEA-U");
+        private final String directory;
+
+        Edition(String directory) {
+            this.directory = directory;
+        }
+
+        public String getDirectory() {
+            return directory;
+        }
+    }
+}
